backend/requirements.txt
fastapi
uvicorn
websockets
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
geoalchemy2
shapely
python-dotenv
alembic
twilio
